Improve Clover based data generation methods
Because of a bug in 'CallMaven' the '-P' argument was always present
regardless of the content of the profiles list parameter. This fix
makes the 'profiles' parameter optional and some refactoring was made
to skip the '-P' if no profiles are provided.

Add self patching and cleanup steps to the coverage generator in
'MutantCode' like it was implemented in the results generator before.

Add a new script that generates coverage and results matrices for
directory based mutants using Clover.

Fix some typos that were related mainly to 'FeedbackModes'.